Rename identity population link_auth flag
Replaced `link_userpass_auth` with `link_auth` across the identity population target config, logs, docs, and fixtures to simplify naming. Also moved `sampleIndices` from the identity population file into the shared identity auth link helper and relocated its tests accordingly, keeping behavior the same while improving helper reuse and test organization.
